requirements.txt
transformers
accelerate
bitsandbytes
datasets
sentence-transformers
faiss-cpu
langchain
huggingface_hub
streamlit
